big refactoring, feature based strucutre for create post #1
