+++ src/app/createpost/_components/DraftSkeleton.tsx
