+++ src/app/createpost/_components/PostForm.tsx
@@ -2,9 +2,10 @@
 import { CreatePostForm, PostDraftType, ProjectDraftType } from "@/types";
 import { Loader2 } from "lucide-react";
 import { useFormContext } from "react-hook-form"
-import { Button } from "../ui/button";
-import { Input } from "../ui/input";
 import { useEffect } from "react";
+import { Label } from "@radix-ui/react-label";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
 
 export default function PostForm({
     onSubmit, 
@@ -15,7 +16,7 @@ export default function PostForm({
     : 
     { 
     onSubmit: () => void, 
-    saveDraft: () => void, 
+    saveDraft: (postForm: any) => void, 
     isSavingDraft: boolean, 
     savedFromDraft?: PostDraftType | ProjectDraftType 
     }) {
@@ -32,13 +33,15 @@ export default function PostForm({
     }, [savedFromDraft, reset]);
 
     return(
-        <form className="grid mt-3 w-full gap-5" onSubmit={handleSubmit(onSubmit)}>
+        <form className="grid mt-3 w-full gap-3 px-3" onSubmit={handleSubmit(onSubmit)}>
+                <Label className="text-primary text-sm" htmlFor="title">*Title</Label>
                 <Input defaultValue={savedFromDraft?.title} {...register('title', {
                     required: "Title is required"
                 })} type="text" id="title" placeholder="Title*" className="h-12 rounded-xl" />
 
                 {errors.title && <span className="text-red-500 text-sm">{errors.title.message}</span>}
 
+                <Label className="text-primary text-sm" htmlFor="content">*Content</Label>
                 <Input defaultValue={savedFromDraft && "content" in savedFromDraft ? savedFromDraft.content : ''} {...register('content', {
                     required: "Content is required"
                 })} type="text" id="title" placeholder="Content*" className="h-20 rounded-xl" />
