+++ src/app/createpost/_components/ProjectForm.tsx
@@ -1,13 +1,14 @@
 import { CreateProjectForm, ProjectDraftType } from "@/types";
 import { useFormContext } from "react-hook-form";
-import { Input } from "../ui/input";
-import { Textarea } from "../ui/textarea";
+import { Input } from "../../../components/ui/input";
+import { Textarea } from "../../../components/ui/textarea";
 import { Loader2 } from "lucide-react";
-import { Button } from "../ui/button";
-import ErrorTooltip from "../reusables/FormErrorTooltip";
+import { Button } from "../../../components/ui/button";
+import ErrorTooltip from "../../../components/reusables/FormErrorTooltip";
 import { useEffect } from "react";
+import { Label } from "@radix-ui/react-label";
 
-export default function ProjectForm({ onSubmit, saveDraft, isSavingDraft, savedFromDraft }: { isSavingDraft?: boolean, saveDraft?: () => void, onSubmit: () => void, savedFromDraft?: ProjectDraftType}){
+export default function ProjectForm({ onSubmit, saveDraft, isSavingDraft, savedFromDraft }: { isSavingDraft?: boolean, saveDraft?: (projectForm: any) => void, onSubmit: () => void, savedFromDraft?: ProjectDraftType}){
     const {
         register,
         handleSubmit,
@@ -31,8 +32,9 @@ export default function ProjectForm({ onSubmit, saveDraft, isSavingDraft, savedF
     }, [savedFromDraft, reset]);
 
     return(
-        <form className="grid mt-3 w-full gap-5" onSubmit={handleSubmit(onSubmit)}>
-            <Input placeholder="Title*" {...register("title", {
+        <form className="grid px-3 mt-3 w-full h-96 overflow-auto gap-2" onSubmit={handleSubmit(onSubmit)}>
+            <Label className="text-primary text-sm" htmlFor="title">*Title</Label>
+            <Input id="title" placeholder="Title*" {...register("title", {
                 required: "Title is required",
                 minLength: {
                     value: 4,
@@ -46,7 +48,8 @@ export default function ProjectForm({ onSubmit, saveDraft, isSavingDraft, savedF
 
             {errors.title && <ErrorTooltip>{errors.title.message}</ErrorTooltip>}
 
-            <Textarea className="h-20" placeholder="Description*" {...register("description", {
+            <Label className="text-primary text-sm" htmlFor="desc">*Description</Label>
+            <Textarea className="h-20" placeholder="Description*" id="desc" {...register("description", {
                 required: "Description is required",
                 minLength: {
                     value: 18,
@@ -60,17 +63,19 @@ export default function ProjectForm({ onSubmit, saveDraft, isSavingDraft, savedF
 
             {errors.description && <ErrorTooltip>{errors.description.message}</ErrorTooltip>}
 
+            <Label className="text-primary text-sm" htmlFor="repo-url">*Repository URL</Label>
+            <Input placeholder="Repository URL* (optional)" id="repo-url" {...register("sourceCodeUrl")} />
 
-            <Input placeholder="Repository URL* (optional)" {...register("sourceCodeUrl")} />
+            <Label className="text-primary text-sm" htmlFor="live-url">*Live URL</Label>
+            <Input placeholder="Live URL* (optional)" id="live-url" {...register("liveDemoUrl")} />
 
-            <Input placeholder="Live URL* (optional)" {...register("liveDemoUrl")} />
+            <Label className="text-primary text-sm" htmlFor="issues">*Issues</Label>
+            <Textarea placeholder="Issues* (optional)" id="issues" {...register("issues")} />
 
-            <Textarea placeholder="Issues* (optional)" {...register("issues")} />
-
-            <div className="flex md:justify-end justify-center gap-3 w-full ">
-                {/* {savedFromDraft === null || undefined ? ( */}
+            <div className="flex md:justify-end justify-center gap-3 mb-5 w-full ">
+                {!savedFromDraft && (
                     <Button className="md:w-32 w-1/2" type="button" onClick={saveDraft} variant={'secondary'} disabled={isFormDisabled || isSavingDraft}>{isSavingDraft ? <Loader2 className="animate-spin" /> : "Save Draft"}</Button>
-                {/* ) : null} */}
+                )}
                 <Button className="md:w-32 w-1/2" disabled={isFormDisabled} type="submit">{isSubmitting ? <Loader2 className="animate-spin"/> : "Submit"}</Button>
             </div>
         </form>
