+++ src/app/createpost/_components/UploadedImagesMap.tsx
