+++ src/helpers/helper.ts
@@ -1,5 +1,3 @@
-import { Images } from "@/types";
-
 export function getPostTypeDetails(type: any | undefined){
     let post_type = '';
     let badge_color = '';
@@ -21,15 +19,3 @@ export function getPostTypeDetails(type: any | undefined){
     }
 }
 
-export function mapImagesToObject(imagesUrls: string[]){
-    if(!imagesUrls) return [];
-
-    const imagesObject = imagesUrls.map((url) => ({ url }));
-    return imagesObject;
-}
-
-export function mapImagesObjectToRawArray(imagesUrls: Images[]) : string[]{
-    if(!imagesUrls) return [];
-    
-    return imagesUrls.map((img) => img.url);
-}
